+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `ParseError`, `GeneratorInfallible` now implements std `Error` trait.
-- `Error` now implements `Clone` trait if GE is `Clone`.
+- `Error` now implements `Clone` trait if `GE` is `Clone`.
 
 ## [0.2.0] - 2022-08-17