+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.2.1] - 2022-8-18
+
 ### Changed
 
 - `ParseError`, `GeneratorInfallible` now implements std `Error` trait.
@@ -45,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/7sDream/tui-markup/compare/v0.2.0..HEAD
+[Unreleased]: https://github.com/7sDream/tui-markup/compare/v0.2.1..HEAD
+[0.2.1]: https://github.com/7sDream/tui-markup/compare/v0.2.0..v0.2.1
 [0.2.0]: https://github.com/7sDream/tui-markup/compare/v0.1.1..v0.2.0
 [0.1.1]: https://github.com/7sDream/tui-markup/compare/v0.1.0..v0.1.1
 [0.1.0]: https://github.com/7sDream/tui-markup/releases/tag/v0.1.0